+++ CHANGELOG.md
@@ -11,6 +11,12 @@ The format is based on [Keep a Changelog], and this project adheres to
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 [bc]: https://github.com/dogmatiq/.github/blob/main/VERSIONING.md#changelogs
 
+## [Unreleased]
+
+### Added
+
+- Added `s3set` package, an S3-backed implementation of `set.BinaryStore`.
+
 ## [0.15.1] - 2026-02-07
 
 ### Added